style: black / part of pylint, better error messages

+++ transform/app/transform/schemas/output/other_rp.py
@@ -20,7 +20,7 @@ other_rp_output_schema = {
     "keywords_tg": "array<string>",
     "language": "array<string>",
     "open_access": "boolean",
-    "popularity": "bigint",
+    "popularity": "int",
     "publication_date": "date",
     "publisher": "string",
     "relations": "array<string>",
