+++ transform/app/transform/schemas/output/publication.py
@@ -18,7 +18,7 @@ publication_output_schema = {
     "keywords_tg": "array<string>",
     "language": "array<string>",
     "open_access": "boolean",
-    "popularity": "bigint",
+    "popularity": "int",
     "publication_date": "date",
     "publisher": "string",
     "relations": "array<string>",
