+++ transform/app/transform/schemas/output/software.py
@@ -19,7 +19,7 @@ software_output_schema = {
     "keywords_tg": "array<string>",
     "language": "array<string>",
     "open_access": "boolean",
-    "popularity": "bigint",
+    "popularity": "int",
     "programming_language": "string",
     "publication_date": "date",
     "publisher": "string",
